Split TripDataManager into TripStore and TripFetcher with permanent Trip identity
Restructure the trip data layer to make stale-Trip-reference bugs
unrepresentable and to separate state from fetch orchestration:

- TripStore: pure synchronous main-thread state. Trip instances are
  permanent (one per tripId for the life of the process), so holding a
  Trip reference across data updates is always safe. Memory is bounded
  by evicting trip *payload* (Trip.clearData) via a composed
  android.util.LruCache, never the instance itself.
- TripFetcher: all async machinery (executor, in-flight dedup,
  ensureSchedule/ensureShape). The schedule/shape backfill that
  previously fired as a side effect inside recordTripsForRouteResponse
  is now an explicit step in RoutePoller.
- Drop TripDataManager.clearAll() from RouteMapController route
  switches: the registry is shared with trip details, trip map, and
  location data screens, and wiping it orphaned their Trip references.
- Extract forEachActiveTrip so TripStore recording and TripFetcher
  backfills walk trips-for-route responses identically.
- Rename lastActiveTripId to vehicleActiveTripId and document it (the
  trip the vehicle most recently reported as active); delete the
  uncalled putLastActiveTripId.
