+++ onebusaway-android/src/google/java/org/onebusaway/android/map/googlemapsv2/tripmap/TripRouteOverlay.kt
@@ -30,7 +30,7 @@ import com.google.android.gms.maps.model.MarkerOptions
 import com.google.android.gms.maps.model.Polyline as MapPolyline
 import java.util.concurrent.TimeUnit
 import org.onebusaway.android.R
-import org.onebusaway.android.extrapolation.data.TripDataManager
+import org.onebusaway.android.extrapolation.data.TripStore
 import org.onebusaway.android.io.elements.ObaTripSchedule
 import org.onebusaway.android.map.googlemapsv2.MapHelpV2
 import org.onebusaway.android.map.googlemapsv2.StampedPolylineFactory
@@ -125,7 +125,7 @@ class TripRouteOverlay(
     }
 
     private fun computeEtaSnippet(arrivalTimeSec: Long, now: Long): String? {
-        val serviceDate = TripDataManager.getServiceDate(tripId) ?: return null
+        val serviceDate = TripStore.getServiceDate(tripId) ?: return null
         val scheduledMs = serviceDate + arrivalTimeSec * 1000
         return if (scheduleDeviation == null) {
             formatEta(
